JustBidIt/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pdfplumber==0.11.0
sqlalchemy>=2.0.46
aiofiles==23.2.1
python-multipart==0.0.9
google-generativeai==0.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
pydantic>=2.9.0
python-dotenv==1.0.1
groq